backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
click==8.1.7
Jinja2==3.1.3
itsdangerous==2.1.2
blinker==1.7.0
flask-cors==4.0.0
supabase==1.2.0
python-dotenv==1.0.1
httpx>=0.23.0,<0.25.0
gotrue>=1.1.0,<1.2.0
postgrest>=0.11.0,<0.12.0
aws-wsgi==0.2.7
boto3==1.26.137
python-jose==3.3.0
requests==2.28.2
langchain-community>=0.0.28,<0.1.0
langchain-core>=0.1.30,<0.2.0
langchain-openai>=0.0.8,<0.1.0
pinecone-client>=3.0.0
openai>=1.14.0,<2.0.0
tiktoken>=0.6.0,<0.7.0
numpy>=1.24.0,<2.0.0
